Adding decimal support to scale field
Adding in expanded decimal support for 'Scale' field. I was working in a
large pixel count file and trying to scale down to get a very specific
screen res on layers with varying bounding boxes and found that things
like 35.65% were only partially supported...

Issue:
The character limit is 4, so '105.5' is stops at '105.'
Fix:
Raise the limit to 8 char max.

Issue:
More than two decimal places seems to make the Generator grumpy.
Fix:
Round to nearest two decimals when get/setting params.

Issue:
Removing percentages with decimals in the layer names is broken. Stops
the name grep at the '.' and renames the layer with the remaining
decimal number instead of the base name
Fix:
Add '.' into the allowed characters in the check for foundScaleSize in
parseName()